Add hexagonal chess board and game logic
This commit introduces the core mechanics and functionality for a hexagonal chess game. Classes for the game, board, cell and piece were created. The game class contains a board which is set to be the Default board. The board class contains the layout and methods to access cells. The cell class keeps track of any active pieces. Each piece has a unique id, a name, and a movement direction. An empty Blazor Server has been added as the main project, but is unmodified.