Add Creative Commons License to the project
A Creative Commons Attribution-ShareAlike 4.0 International license has been added to the Hexagonal Chess project. This license was selected as it allows others to remix, adapt, and build upon the project even for commercial purposes, as long as they credit the original creation and license their new creations under the identical terms. Adding this license provides clear terms of distribution and use for other contributors and users of this project.